Refactor PKWT to use divisi_id and jabatan_id relations
Updated PKWT model, migration, controller, and related views to use foreign keys divisi_id and jabatan_id instead of string fields for divisi and jabatan. Added Eloquent relationships for Divisi and JabatanPKWT in the PKWT model. Refactored the form and table UI to use the new structure and improved the display of worker data in the unit detail view.