Add unit-user relationship and update navbar for PIC role
Introduced belongsToMany relationships between User and Unit models via the pic_unit pivot table. Updated the navbar to show a dropdown of units for users with the 'pic' role, linking to a new unit detail route. Added the corresponding route and fixed a cancel button link in the unit edit view.